脚本修复

+++ run.sh
@@ -78,10 +78,18 @@ restart(){
 #下载
 download(){
   echo "文件ID为：$1"
-  fileUrl = `curl -X POST "http://localhost:8888/v1/LuaApiCaller?qq=3022645754&funcname=OidbSvc.0x6d6_2" -H "accept: */*" -H "Content-Type: application/json" -d "{ \"GroupID\": 901158551, \"FileID\": \"$1\"}"`
+  fileJson=`curl -X POST "http://localhost:8888/v1/LuaApiCaller?qq=3022645754&funcname=OidbSvc.0x6d6_2" -H "accept: */*" -H "Content-Type: application/json" -d "{ \"GroupID\": 901158551, \"FileID\": \"$1\"}"`
+  fileJson1=${fileJson#*Url\":\"}
+  fileUrl=${fileJson1%\"*}
   echo "获取到文件Url为${fileUrl}"
-  mv ${JAR_NAME} "${JAR_NAME}.temple"
+  if [ ! -f "${JAR_NAME}.temple" ]; then
+    mv ${JAR_NAME} "${JAR_NAME}.temple"
+  else
+    rm -f "${JAR_NAME}.temple"
+    mv ${JAR_NAME} "${JAR_NAME}.temple"
+  fi
   wget -c ${fileUrl} -O ${JAR_NAME}
+  restart
 }
 
 ##更新
